clientes/requirements.txt
googlemaps==4.10.0

boto3==1.24.92
botocore==1.27.92

channels==3.0.5
channels-redis==3.4.1

confluent-kafka==2.1.1

Django==3.2.16
djangorestframework-api-response==0.0.30

django-cors-headers==3.13.0
django-environ==0.9.0
django-redis==5.2.0

django-storages==1.13.1
jsonpickle==3.0.1

djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0

social-auth-core==4.3.0

Pillow==9.2.0
psycopg2==2.9.4

python-slugify==6.1.2
PyJWT==2.5.0


web3==5.31.3
eth-account==0.5.9

whitenoise==6.2.0
uvicorn==0.20.0

rsa==4.9
LWE4==0.0.0.2

django-location-field==2.7.1
django-map-widgets==0.1.3